Move out `struct fsm_allocator` to its own header.
